third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/script-snippet-model-expected.txt

Tests script snippet model.


Running: testCreateEditRenameRemove
UISourceCodeAdded: Script snippet #1
Snippet content: 
Snippet content set.
Snippet content: <snippet content>
Snippet1 created.
UISourceCodeAdded: Script snippet #2
Snippet2 created.
Renaming snippet to 'foo' ...
Snippet renamed successfully.
UISourceCode name is 'foo' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Renaming snippet to '   ' ...
Snippet was not renamed.
UISourceCode name is 'foo' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Renaming snippet to ' bar ' ...
Snippet renamed successfully.
UISourceCode name is 'bar' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Renaming snippet to 'foo' ...
Snippet renamed successfully.
UISourceCode name is 'foo' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Renaming snippet to 'bar' ...
Snippet renamed successfully.
UISourceCode name is 'bar' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Renaming snippet to 'foo' ...
Snippet was not renamed.
UISourceCode name is 'bar' now.
Number of uiSourceCodes in workspace: 2
Number of snippets in the storage: 2
Snippet content: <snippet content>
UISourceCodeRemoved: foo
UISourceCodeRemoved: bar
UISourceCodeAdded: Script snippet #3
Snippet3 created.
UISourceCodeRemoved: Script snippet #3
Snippet3 deleted.
Number of uiSourceCodes in workspace: 0
Number of snippets in the storage: 0

Running: testEvaluate
Last evaluation source url for snippet: snippets:///1_1
Snippet execution result: undefined
Last evaluation source url for snippet: snippets:///2_2
Snippet execution result: function doesNothing() {
    var  i = 2+2;
}
Last evaluation source url for snippet: snippets:///1_3
Snippet execution result: undefined
Last evaluation source url for snippet: snippets:///3_4
Snippet execution result: 1

Running: testEvaluateEditReload
Snippet execution result: undefined
Page reloaded.

Running: testEvaluateInWorker
Last evaluation source url for snippet: snippets:///1_1
Snippet execution result: 4

Running: testDangerousNames
